Added preliminary support for Chest SRM - very wonky
This script as originally designed by Savestate was meant to find facing/movement angles, so I had to do a bit of a hack to get it to find setups for collision angles. Essentially, I compute all the possible working collision angles, then back movement angles out of that, and let the script find movement angles, since that's what it's good at. As a result you will not be notified of which wall you need to collide with to get the collision angle, but with basic addition/subtraction it should be pretty obvious for Observatory at least, since we only have cardinal direction walls (multiples of 0x4000). Example: you're given a setup for FD mask that creates a movement angle of 0x75ab. But you know FD mask is item 0x35 on English. So you'll need to collide with wall 0x4000 to get the correct collision angle (since 0x75 - 0x35 = 0x40).